commit 7b2eb3f3c068e560c7357b0f314172448bb8d1e0
    12057 Writing part of the string to stderr makes zlogin exit
commit d2dd27964b9dfc03118548f9509e4706787c2a69
    12088 Cannot build iasl with bison 3.5 [-Werror=char-subscripts]
commit bf74bfd433cc5106d2fffc63678efb6ba4a3694c
    12090 loader.efi: efi_readkey_ex needs to check the key despite the shift status or toggle status
commit 5947648b7f5c085635051e1f7aa083a309542467
    12069 Backport sh_delay() and tvsleep() from ksh-2020.0.0
commit fc5c75cf5edb072564020725faa0c4313714f09f
    12051 re-enable ZFS trim by default
commit 8b35e52344673c75ba6a446ced1fb5c36b52a242
    12070 sata SSDs attached to sata ports can't trim
commit 2ed5f78a049996104f9dcce38d0c0c1735dd0e7a
    12091 loader: biosdisk.c: Use symbolic names for int13 calls
commit bf6cb86ec437546144857d9aa94ef222ec6763c0
    12082 libpctx: cast between incompatible function types
commit c94f4b0313ed735fc39e1b15e1fa48c0e6e3730f
    12081 libldap5: cast between incompatible function types
commit 3c19a1689122901345a6089d081aa2de4a1096da
    12079 audit_plugins: cast between incompatible function types
commit adbb29bd77dbfbbf129b5e3bd6ccc5bd4074fd79
    12073 loader: devopen dereference after free
commit fdf04373777e703cebbbce0f774ae970eb6030f7
    12072 loader: vdisk dereference after free

+++ usr/src/lib/libldap5/sources/ldap/common/open.c
@@ -34,7 +34,7 @@
  */
 
 #if 0
-#ifndef lint 
+#ifndef lint
 static char copyright[] = "@(#) Copyright (c) 1995 Regents of the University of Michigan.\nAll rights reserved.\n";
 #endif
 #endif
@@ -82,7 +82,7 @@ struct nsldapi_ldap_error {
 #else
 __declspec ( thread ) int	nsldapi_gldaperrno;
 __declspec ( thread ) char	*nsldapi_gmatched = NULL;
-__declspec ( thread ) char	*nsldapi_gldaperror = NULL; 
+__declspec ( thread ) char	*nsldapi_gldaperror = NULL;
 #endif /* _WINDOWS */
 
 #ifdef _WINDOWS
@@ -245,18 +245,18 @@ get_ld_error( char **matched, char **errmsg, void *dummy )
 
         le = pthread_getspecific( nsldapi_key );
 	if (le != NULL) {
-        	if ( matched != NULL ) {
-                	*matched = le->le_matched;
-        	}
-        	if ( errmsg != NULL ) {
-                	*errmsg = le->le_errmsg;
-        	}
-        	return( le->le_errno );
+		if ( matched != NULL ) {
+			*matched = le->le_matched;
+		}
+		if ( errmsg != NULL ) {
+			*errmsg = le->le_errmsg;
+		}
+		return( le->le_errno );
 	} else {
-        	if ( matched != NULL )
-                	*matched = NULL;
-        	if ( errmsg != NULL )
-                	*errmsg = NULL;
+		if ( matched != NULL )
+			*matched = NULL;
+		if ( errmsg != NULL )
+			*errmsg = NULL;
 	}
 	return (LDAP_SUCCESS);
 }
@@ -292,7 +292,7 @@ static struct ldap_extra_thread_fns
 #ifdef _WINDOWS
 		0
 #else
-		(void *(*)(void))pthread_self
+		(void *(*)(void))(uintptr_t)pthread_self
 #endif /* _WINDOWS */
 		};
 
@@ -393,8 +393,8 @@ ldap_version( LDAPVersion *ver )
 		ver->sdk_version = (int)(VI_PRODUCTVERSION * 100);
 		ver->protocol_version = LDAP_VERSION_MAX * 100;
 		ver->SSL_version = SSL_VERSION * 100;
-		/* 
-		 * set security to none by default 
+		/*
+		 * set security to none by default
 		 */
 
 		ver->security_level = LDAP_SECURITY_NONE;
@@ -710,7 +710,7 @@ ldap_x_hostlist_next( char **hostp, int *portp,
 		status->lhs_nexthost = NULL;
 	}
 
-	/* 
+	/*
 	 * Look for closing ']' and skip past it before looking for port.
 	 */
 	if ( squarebrackets && NULL != ( q = strchr( *hostp, ']' ))) {
